Extract getFullRect into its own module
The logic around this is complicated and isolated enough to warrant
being in a module of its own.
